working version with internal client
