+++ enterprise/internal/insights/aggregation/limited_aggregator.go
@@ -1,4 +1,4 @@
-package streaming
+package aggregation
 
 import (
 	"sort"
