+++ enterprise/internal/insights/aggregation/limited_aggregator_test.go
@@ -1,4 +1,4 @@
-package streaming
+package aggregation
 
 import (
 	"testing"
